Stub broadcast channel for unspported contexts
