+++ src/lib/broadcast/index.ts
@@ -1,11 +1,2 @@
-export default class BroadcastChannel {
-  constructor(public name: string) {}
-  postMessage(_data: any) {}
-  close() {}
-  onmessage: (event: MessageEvent) => void = () => {}
-  addEventListener(_type: string, _listener: (event: MessageEvent) => void) {}
-  removeEventListener(
-    _type: string,
-    _listener: (event: MessageEvent) => void,
-  ) {}
-}
+import Stub from '#/lib/broadcast/stub'
+export default Stub
