+++ src/lib/broadcast/index.web.ts
@@ -1,1 +1,2 @@
-export default BroadcastChannel
+import Stub from '#/lib/broadcast/stub'
+export default 'BroadcastChannel' in window ? window.BroadcastChannel : Stub
